refactor(linter): move C100 array policy into facts
