+++ crates/shuck-linter/src/facts/builder.rs
@@ -861,6 +861,7 @@ impl<'a, 'analysis> LinterFactsBuilder<'a, 'analysis> {
             semantic_artifacts: self.semantic_artifacts,
             source,
             line_index: self._indexer.line_index(),
+            shell: self.shell,
             commands,
             command_fact_indices_by_id,
             structural_command_ids,
@@ -892,6 +893,7 @@ impl<'a, 'analysis> LinterFactsBuilder<'a, 'analysis> {
             possible_variable_misspelling_use_scan: OnceLock::new(),
             possible_variable_misspelling_index: OnceLock::new(),
             possible_variable_misspelling_scope_compat_name_uses: OnceLock::new(),
+            plain_unindexed_array_references: OnceLock::new(),
             redundant_echo_space_facts: OnceLock::new(),
             suppressed_subscript_reference_spans,
             subscript_later_suppression_reference_spans,
